// web/models/datasets.ts
export type DatasetPermission = 'only_me' | 'all_team_members' | 'partial_members'